+++ RocketWanderer/WpfApp/Views/Screens/RecordsTableViewWpf.cs
@@ -111,8 +111,8 @@ namespace WpfApp.Views.Screens
       _scrollViewer.Width = tableSize.X;
       _scrollViewer.Height = tableSize.Y;
 
-      int leftOffset = parParentSize.X / 2 - tableSize.X / 2;
-      int topOffset = parParentSize.Y / 2 - tableSize.Y / 2 + 40;
+      double leftOffset = parParentSize.X / 2 - tableSize.X / 2;
+      double topOffset = parParentSize.Y / 2 - tableSize.Y / 2 + 40;
 
       System.Windows.Controls.Canvas.SetLeft(_scrollViewer, leftOffset);
       System.Windows.Controls.Canvas.SetTop(_scrollViewer, topOffset);
